ct: remove typhoontimer
